+++ lib/discard/model.rb
@@ -50,6 +50,25 @@ module Discard
         kept.each(&:discard)
       end
 
+      # Discards the records by instantiating each
+      # record and calling its {#discard!} method.
+      # Each object's callbacks are executed.
+      # Returns the collection of objects that were discarded.
+      #
+      # Note: Instantiation, callback execution, and update of each
+      # record can be time consuming when you're discarding many records at
+      # once. It generates at least one SQL +UPDATE+ query per record (or
+      # possibly more, to enforce your callbacks). If you want to discard many
+      # rows quickly, without concern for their associations or callbacks, use
+      # #update_all!(discarded_at: Time.current) instead.
+      #
+      # ==== Examples
+      #
+      #   Person.where(age: 0..18).discard_all!
+      def discard_all!
+        kept.each(&:discard!)
+      end
+
       # Undiscards the records by instantiating each
       # record and calling its {#undiscard} method.
       # Each object's callbacks are executed.
@@ -68,6 +87,25 @@ module Discard
       def undiscard_all
         discarded.each(&:undiscard)
       end
+
+      # Undiscards the records by instantiating each
+      # record and calling its {#undiscard!} method.
+      # Each object's callbacks are executed.
+      # Returns the collection of objects that were undiscarded.
+      #
+      # Note: Instantiation, callback execution, and update of each
+      # record can be time consuming when you're undiscarding many records at
+      # once. It generates at least one SQL +UPDATE+ query per record (or
+      # possibly more, to enforce your callbacks). If you want to undiscard many
+      # rows quickly, without concern for their associations or callbacks, use
+      # #update_all!(discarded_at: nil) instead.
+      #
+      # ==== Examples
+      #
+      #   Person.where(age: 0..18).undiscard_all!
+      def undiscard_all!
+        discarded.each(&:undiscard!)
+      end
     end
 
     # @return [Boolean] true if this record has been discarded, otherwise false
@@ -85,11 +123,17 @@ module Discard
       !archived? && !discarded?
     end
 
+    # @return [Boolean] false if this record has been discarded, otherwise true
+    def undiscarded?
+      !discarded?
+    end
+    alias kept? undiscarded?
+
     # Discard the record in the database
     #
     # @return [Boolean] true if successful, otherwise false
     def discard
-      return if discarded?
+      return false if discarded?
       run_callbacks(:discard) do
         update_attributes("#{self.class.discard_column}": Time.current, "#{self.class.archive_column}": nil)
       end