+++ spec/discard/model_spec.rb
@@ -37,6 +37,14 @@ RSpec.describe Discard::Model do
         expect(post).not_to be_discarded
       end
 
+      it "should be undiscarded?" do
+        expect(post).to be_undiscarded
+      end
+
+      it "should be kept?" do
+        expect(post).to be_kept
+      end
+
       describe '#discard' do
         it "sets discarded_at" do
           expect {
@@ -109,6 +117,14 @@ RSpec.describe Discard::Model do
 
       it "should be discarded?" do
         expect(post).to be_discarded
+      end
+
+      it "should not be undiscarded?" do
+        expect(post).to_not be_undiscarded
+      end
+
+      it "should not be kept?" do
+        expect(post).to_not be_kept
       end
 
       describe '#discard' do
@@ -269,6 +285,14 @@ RSpec.describe Discard::Model do
         expect(post).not_to be_discarded
       end
 
+      it "should be undiscarded?" do
+        expect(post).to be_undiscarded
+      end
+
+      it "should be kept?" do
+        expect(post).to be_kept
+      end
+
       describe '#discard' do
         it "sets discarded_at" do
           expect {
@@ -319,6 +343,14 @@ RSpec.describe Discard::Model do
 
       it "should be discarded?" do
         expect(post).to be_discarded
+      end
+
+      it "should not be undiscarded?" do
+        expect(post).to_not be_undiscarded
+      end
+
+      it "should not be kept?" do
+        expect(post).to_not be_kept
       end
 
       describe '#discard' do
@@ -416,13 +448,44 @@ RSpec.describe Discard::Model do
         user2 = User.create!
 
         2.times { user1.comments.create! }
-        2.times { user1.comments.create! }
+        2.times { user2.comments.create! }
 
         user1.comments.discard_all
-
-        expect(user1.comments).to all(be_discarded)
-        expect(user2.comments).to all(be_undiscarded)
+        user1.comments.each do |comment|
+          expect(comment).to be_discarded
+          expect(comment).to_not be_undiscarded
+          expect(comment).to_not be_kept
+        end
+        user2.comments.each do |comment|
+          expect(comment).to_not be_discarded
+          expect(comment).to be_undiscarded
+          expect(comment).to be_kept
+        end
       end
+    end
+  end
+
+  describe '.discard_all!' do
+    with_model :Post, scope: :all do
+      table do |t|
+        t.string :title
+        t.datetime :discarded_at
+        t.timestamps null: false
+      end
+
+      model do
+        include Discard::Model
+      end
+    end
+
+    let!(:post) { Post.create!(title: "My very first post") }
+    let!(:post2) { Post.create!(title: "A second post") }
+
+    it "can discard all posts" do
+      expect {
+        Post.discard_all!
+      }.to   change { post.reload.discarded? }.to(true)
+        .and change { post2.reload.discarded? }.to(true)
     end
   end
 
@@ -459,6 +522,30 @@ RSpec.describe Discard::Model do
       Post.where(id: []).undiscard_all
       expect(post.reload).to be_discarded
       expect(post2.reload).to be_discarded
+    end
+  end
+
+  describe '.undiscard_all!' do
+    with_model :Post, scope: :all do
+      table do |t|
+        t.string :title
+        t.datetime :discarded_at
+        t.timestamps null: false
+      end
+
+      model do
+        include Discard::Model
+      end
+    end
+
+    let!(:post) { Post.create!(title: "My very first post", discarded_at: Time.now) }
+    let!(:post2) { Post.create!(title: "A second post", discarded_at: Time.now) }
+
+    it "can undiscard all posts" do
+      expect {
+        Post.undiscard_all!
+      }.to   change { post.reload.discarded? }.to(false)
+        .and change { post2.reload.discarded? }.to(false)
     end
   end