+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 ### Unreleased
-* enhancements
-  * `discard_all` and `undiscard_all` now return affected records
-  * Add `discard!` and `undiscard!`
-  
+
+### Version 1.2.0
+Release date: 2020-02-17
+
+* Add `discard_all!` and `undiscard_all!`
+* Add `undiscarded?` and `kept?` to match the scopes of the same names
+
+### Version 1.1.0
+Release date: 2019-05-03
+
+* Support for ActiveRecord 6
+* `discard_all` and `undiscard_all` now return affected records
+* Add `discard!` and `undiscard!`
+
 ### Version 1.0.0
 Release date: 2018-03-16